Fix: Tampilan

+++ app/Http/Controllers/SupplierController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Alternatif;
-use App\Models\Kriteria;
+use App\Models\Produk;
 use App\Models\Kota;
 use App\Models\Kecamatan;
 use App\Models\Supplier;
@@ -71,10 +71,15 @@ class SupplierController extends Controller {
     return redirect('admin/supplier')->with('success', 'Berhasil mengubah data.');
   }
 
-  public function delete($id_supplier) {
-    $data = Supplier::find($id_supplier);
-    $data->delete();
-    return redirect('admin/supplier')->with('success', 'Berhasil menghapus data.');
+  public function delete($supplier_id ) {
+    $data = Supplier::find($supplier_id);
+    $data_satuan_produk = Produk::where('supplier_id', $supplier_id)->count();
+    if($data_satuan_produk > 0) {
+      return redirect('admin/supplier')->with('error', 'Tidak bisa menghapus data! Masih ada Produk yang berasal dari '.$data->nama);
+    } else {
+      $data->delete();
+      return redirect('admin/supplier')->with('success', 'Berhasil menghapus data.');
+    }
   }
 
 }
